src/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["woojin-an <dkswoojin@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {version = "^0.115.0", extras = ["standard"]}
uvicorn = "^0.32.0"
gunicorn = "^20.1.0"
pydantic-settings = "^2.6.1"
tortoise-orm = {version = "^0.21.6", extras = ["aiomysql"]}
aerich = "^0.7.2"
cryptography = "^43.0.3"
greenlet = "^3.1.1"
python-multipart = "^0.0.17"
python-dotenv = "^1.0.1"
httpx = "^0.27.2"
redis = "^5.2.0"
boto3 = "^1.35.57"
pillow = "^11.0.0"
aiomysql = "^0.2.0"
boto3-stubs = "^1.35.67"
botocore-stubs = "^1.35.67"
bcrypt = "^4.2.1"
pyjwt = "^2.10.0"

[tool.poetry.group.dev.dependencies]
black = "^24.10.0"
isort = "^5.13.2"
mypy = "^1.13.0"
coverage = "^7.6.4"
pytest = "^8.3.3"
tortoise-orm = {extras = ["mypy"], version = "^0.21.7"}
types-pillow = "^10.2.0.20240822"

[tool.aerich]
tortoise_orm = "core.database.db_settings.TORTOISE_ORM"
location = "./migrations"
src_folder = "./"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
python_version = "3.12"
strict = true
ignore_missing_imports = true
#exclude = "tests/"
plugins = ['pydantic.mypy']

[[tool.mypy.overrides]]
module = "PIL.*"
ignore_missing_imports = true

[tool.black]
line-length = 120

[tool.isort]
profile = "black"
line_length = 120
#
#[tool.pytest.ini_options]
#asyncio_mode = "auto"
#asyncio_fixture_loop_scope = "function"
#markers = [
#    "asyncio: mark test as async",
#]
#testpaths = ["tests"]
#python_files = ["test_*.py"]
#python_classes = ["Test*"]
#python_functions = ["test_*"]